planventure-api/requirements.txt
# Core dependencies
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0

# JWT authentication
flask-jwt-extended==4.5.3
werkzeug==2.3.7

# Password hashing
bcrypt==4.0.1

# Database migration
flask-migrate==4.0.5

# API documentation
flasgger==0.9.7.1
flask-swagger-ui==4.11.1

# Testing dependencies
requests==2.31.0